# perf: move screen focus to LHNOptionsList to avoid re-renders