--- a/src/hooks/useIsScreenFocus.ts
+++ b/src/hooks/useIsScreenFocus.ts
@@ -1,0 +1,17 @@
+import {useFocusEffect} from '@react-navigation/native';
+import {useCallback, useState} from 'react';
+
+export default function useIsScreenFocusedRef() {
+    const [isScreenFocused, setIsScreenFocused] = useState(false);
+
+    useFocusEffect(
+        useCallback(() => {
+            setIsScreenFocused(true);
+            return () => {
+                setIsScreenFocused(false);
+            };
+        }, []),
+    );
+
+    return isScreenFocused;
+}