--- a/src/hooks/useIsScreenFocus.ts
+++ b/src/hooks/useIsScreenFocus.ts
@@ -1,7 +1,7 @@
 import {useFocusEffect} from '@react-navigation/native';
 import {useCallback, useState} from 'react';
 
-export default function useIsScreenFocusedRef() {
+export default function useIsScreenFocused() {
     const [isScreenFocused, setIsScreenFocused] = useState(false);
 
     useFocusEffect(